# server/openapi.yaml
openapi: 3.1.0
info:
  title: Image Annotation Hybrid Project API
  version: 1.0.0
  description: |
    Remote API for hybrid image-annotation projects. Mutating requests are
    authenticated, authorized per project, audited, and exposed through the
    incremental change stream.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/projects/{projectId}/issues:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Issues]
      operationId: listIssues
      responses:
        "200":
          description: Project issues
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Issues]
      operationId: createIssue
      responses:
        "200":
          description: Created issue
        default:
          $ref: "#/components/responses/Error"
  /api/v1/issues/{issueId}:
    parameters:
      - $ref: "#/components/parameters/IssueId"
    get:
      tags: [Issues]
      operationId: getIssue
      responses:
        "200":
          description: Issue detail
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Issues]
      operationId: updateIssue
      responses:
        "200":
          description: Updated issue
        default:
          $ref: "#/components/responses/Error"
  /api/v1/issues/{issueId}/transition:
    parameters:
      - $ref: "#/components/parameters/IssueId"
    post:
      tags: [Issues]
      operationId: transitionIssue
      responses:
        "200":
          description: Transitioned issue
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IssueId:
      name: issueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: [string, "null"]
        details:
          type: [object, "null"]
          additionalProperties: true